Remove failed sockets
KDCProxy appears to rely on Windows-style select() behavior, where a
failed connection will be listed in the exception set.  On Linux, this
results in a failed connection not being removed.  Instead, kdcproxy
will keep trying and failing to send data in a loop to a socket for some
seconds until it times out, logging errors all the while.

Fix.